[MOL-20717][RP] split variants into separate files.
