+++ src/language-switcher/types.ts
@@ -15,9 +15,19 @@ export interface LanguageSwitcherProps
     "data-testid"?: string | undefined;
 }
 
+export interface VariantInternalProps {
+    selectedLanguage: LanguageCode;
+    onSelectLanguage: (code: LanguageCode) => void;
+    testId: string;
+    otherProps: React.HTMLAttributes<HTMLDivElement>;
+}
+
 // =============================================================================
 // CONSTANTS
 // =============================================================================
+export const ARIA_LABEL =
+    "Choose language / 选择语言 / Pilih bahasa / மொழியை தேர்ந்தெடுக்கவும்";
+
 export const LANGUAGE_DISPLAY_MAP: Record<LanguageCode, string> = {
     en: "English",
     zh: "中文",
